requirements.txt
accelerate==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
datasets==4.7.0
dill==0.4.0
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2025.12.0
gradio==6.9.0
gradio_client==2.3.0
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
numpy==2.2.6
orjson==3.11.7
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.0
rank-bm25==0.2.2
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.15.2
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==4.39.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
xxhash==3.6.0
yarl==1.23.0
